@@ -1,33 +1,57 @@
 module github.com/VuThanhThien/golang-gorm-postgres/authentication
 
-go 1.21
+go 1.22.0
 
-toolchain go1.22.5
+toolchain go1.23.0
 
 require (
-	github.com/gin-contrib/cors v1.5.0
-	github.com/gin-gonic/gin v1.9.1
+	github.com/ahmetb/go-linq/v3 v3.2.0
+	github.com/cenkalti/backoff/v4 v4.3.0
+	github.com/gin-contrib/cors v1.7.2
+	github.com/gin-gonic/gin v1.10.0
 	github.com/golang-jwt/jwt v3.2.2+incompatible
 	github.com/google/uuid v1.6.0
-	github.com/spf13/viper v1.18.2
-	golang.org/x/crypto v0.26.0
-	golang.org/x/time v0.5.0
-	gorm.io/driver/postgres v1.5.6
-	gorm.io/gorm v1.25.7
+	github.com/grpc-ecosystem/grpc-gateway/v2 v2.22.0
+	github.com/hashicorp/consul/api v1.29.4
+	github.com/iancoleman/strcase v0.3.0
+	github.com/pkg/errors v0.9.1
+	github.com/satori/go.uuid v1.2.0
+	github.com/sirupsen/logrus v1.9.3
+	github.com/spf13/viper v1.19.0
+	github.com/streadway/amqp v1.1.0
+	golang.org/x/crypto v0.27.0
+	golang.org/x/time v0.6.0
+	google.golang.org/grpc v1.67.0
+	google.golang.org/grpc/cmd/protoc-gen-go-grpc v1.5.1
+	gorm.io/driver/postgres v1.5.9
+	gorm.io/gorm v1.25.12
 )
 
 require (
-	github.com/bytedance/sonic/loader v0.1.1 // indirect
+	github.com/armon/go-metrics v0.4.1 // indirect
+	github.com/bytedance/sonic/loader v0.2.0 // indirect
 	github.com/cloudwego/base64x v0.1.4 // indirect
 	github.com/cloudwego/iasm v0.2.0 // indirect
+	github.com/fatih/color v1.17.0 // indirect
+	github.com/hashicorp/errwrap v1.1.0 // indirect
+	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
+	github.com/hashicorp/go-hclog v1.6.3 // indirect
+	github.com/hashicorp/go-immutable-radix v1.3.1 // indirect
+	github.com/hashicorp/go-multierror v1.1.1 // indirect
+	github.com/hashicorp/go-rootcerts v1.0.2 // indirect
+	github.com/hashicorp/golang-lru v1.0.2 // indirect
+	github.com/hashicorp/serf v0.10.1 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mitchellh/go-homedir v1.1.0 // indirect
+	google.golang.org/genproto/googleapis/api v0.0.0-20240924160255-9d4c2d233b61 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240924160255-9d4c2d233b61 // indirect
 )
 
 require (
 	github.com/KyleBanks/depth v1.2.1 // indirect
-	github.com/bytedance/sonic v1.11.6 // indirect
+	github.com/bytedance/sonic v1.12.3 // indirect
 	github.com/fsnotify/fsnotify v1.7.0 // indirect
-	github.com/gabriel-vasile/mimetype v1.4.3 // indirect
+	github.com/gabriel-vasile/mimetype v1.4.5 // indirect
 	github.com/gin-contrib/gzip v1.0.1
 	github.com/gin-contrib/sse v0.1.0 // indirect
 	github.com/go-openapi/jsonpointer v0.21.0 // indirect
