@@ -1,8 +1,13 @@
 package models
 
+import (
+	"gorm.io/gorm"
+)
+
 type Merchant struct {
-	Base
+	gorm.Model
 	Name        string `gorm:"type:varchar(255);not null"`
 	MerchantId  string `gorm:"uniqueIndex;not null"`
 	Description string `gorm:"type:varchar(255)"`
+	UserId      uint   `gorm:"not null"`
 }
